Table generator iteration
Add ParamIter and framework for specifying the order in which
simulation parameter values are incremented (in odometric fashion),
when multiple iterations of simulation parameters are specified.

The simulator returns parallelization information as a value, so the
output of the simulator can eventually be HTML or Markdown tables.

SQUASH later commits into here.

+++ go/scheduler/alg/scheduler_test.go
@@ -73,7 +73,7 @@ func (tss *TransactionSliceSource) GetTransactions() []*Transaction {
 			}
 			ts[ix].WriteSet.Add(loc)
 		}
-		ts[ix].CreationSeqno = uint(tss.nextTid)
+		ts[ix].CreationSeqno = tss.nextTid
 		tss.nextTid++
 	}
 	return ts
